dev_spec_Alexandre

+++ DEV_SPEC/Alexandre/dev_modflow_hs1d/piezometry.py
@@ -15,10 +15,11 @@ class extract:
 		data_folder = out_path + 'data/piezometric/'
 		if not os.path.exists(data_folder):
 				os.makedirs(data_folder)	
-		#self.download_init_data(data_folder)
-		self.exctract_piezos_from_watershed(data_folder, geographic)
-		#self.extract_data_from_code_bss(data_folder)
-		self.load_piezometric_data(data_folder)
+		self.download_init_data(data_folder)
+		piezos = self.exctract_piezos_from_watershed(data_folder, geographic)
+		if os.path.exists(data_folder + 'shapefile/piezos.shp'):
+			self.extract_data_from_code_bss(data_folder)
+			self.load_piezometric_data(data_folder)
 
 	def download_init_data(self,data_folder):
 		filename = data_folder + 'piezometers.zip'
@@ -37,35 +38,41 @@ class extract:
 		piezos_fr = gpd.read_file(piezos_shp)
 		piezos_fr.to_crs(epsg=2154, inplace=True)
 		piezos = gpd.clip(piezos_fr, watershed)
-		piezos.to_file(data_folder + 'shapefile/piezos.shp')
-		self.codes_bss = piezos['code_bss'].tolist()
-		for i in range (0, len(self.codes_bss)):
-			self.codes_bss[i] = self.codes_bss[i].replace('/','_')
-		self.x_coord = piezos['geometry'].x.tolist()
-		self.y_coord = piezos['geometry'].y.tolist()
+		if len(piezos)!=0:
+			piezos.to_file(data_folder + 'shapefile/piezos.shp')
+			self.codes_bss = piezos['code_bss'].tolist()
+			for i in range (0, len(self.codes_bss)):
+				self.codes_bss[i] = self.codes_bss[i].replace('/','_')
+			self.x_coord = piezos['geometry'].x.tolist()
+			self.y_coord = piezos['geometry'].y.tolist()
+		return self, piezos
 
 	def extract_data_from_code_bss(self,data_folder):
 		for code in self.codes_bss:
 			code_ = code.replace('_','/')
-			url = 'https://ades.eaufrance.fr/Fiche/PtEau?Code=' + code_
-			chrome_options = webdriver.ChromeOptions()
-			prefs = {'download.default_directory' : data_folder.replace('/','\\')}
-			chrome_options.add_experimental_option('prefs', prefs)
-			driver = webdriver.Chrome(chrome_options=chrome_options)
-			driver.get(url)
-			elem = driver.find_element_by_link_text('Tout télécharger')
-			elem.click()
-			compt = 0
-			while (compt==0):
-				if len(glob.glob(data_folder + '*.zip')) == 1:
-					compt +=1
-					time.sleep(1)
-				time.sleep(1)
-			driver.close()
-			file = glob.glob(data_folder+'/*.zip')[0]
-			with zipfile.ZipFile(file, 'r') as zip_ref:
-				zip_ref.extractall(data_folder+'/'+code)
-			os.remove(file)
+			if not os.path.exists(data_folder+'/'+code):
+				url = 'https://ades.eaufrance.fr/Fiche/PtEau?Code=' + code_
+				chrome_options = webdriver.ChromeOptions()
+				prefs = {'download.default_directory' : data_folder.replace('/','\\')}
+				chrome_options.add_experimental_option('prefs', prefs)
+				driver = webdriver.Chrome(chrome_options=chrome_options)
+				driver.get(url)
+				try:
+					elem = driver.find_element_by_link_text('Tout télécharger')
+					elem.click()
+					compt = 0
+					while (compt==0):
+						if len(glob.glob(data_folder + '*.zip')) == 1:
+							compt +=1
+							time.sleep(1)
+						time.sleep(1)
+					driver.close()
+					file = glob.glob(data_folder+'/*.zip')[0]
+					with zipfile.ZipFile(file, 'r') as zip_ref:
+						zip_ref.extractall(data_folder+'/'+code)
+					os.remove(file)
+				except:
+					self.codes_bss.remove(code)
 
 	def load_piezometric_data(self,data_folder):
 		self.depth = pd.DataFrame()
